pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "aiofiles>=25.1.0",
    "anthropic[bedrock]>=0.75.0",
    "ariadne-codegen>=0.17.0",
    "beanie>=2.0.1",
    "bedrock-agentcore>=1.2.0",
    "boto3-stubs[s3]>=1.42.2",
    "ddgs>=9.10.0",
    "docling>=2.71.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.128.7",
    "fastmcp>=2.14.4",
    "firecrawl-py>=4.9.0",
    "google-search-results>=2.4.2",
    "langchain>=1.1.0",
    "langchain-aws>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-exa>=1.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.0",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "langgraph-api>=0.6.35",
    "langgraph-checkpoint-postgres>=3.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langmem>=0.0.30",
    "langsmith>=0.4.49",
    "llama-index>=0.14.13",
    "llama-index-llms-openai>=0.6.16",
    "llama-index-tools-mcp>=0.4.5",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic>=2.0.0",
    "pymongo>=4.10",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "semanticscholar>=0.11.0",
    "taskiq>=0.12.1",
    "taskiq-aio-pika>=0.5.0",
    "taskiq-redis>=1.2.2",
    "tavily-python>=0.7.13",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
    "wikipedia>=1.4.0",
    "xmltodict>=1.0.2",
    "yfinance>=1.1.0",
]

# Dev-only tools and test deps
[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "black",
    "mypy", 
    "pyright",
    "pydantic>=2.0.0",
    "ipykernel",
    "ariadne-codegen[subscriptions]>=0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 

 [tool.hatch.build.targets.wheel]
 packages = ["src/research_agent"]

[tool.ruff]
# Your actual source tree
src = ["src"]
line-length = 100
# Tools are still catching up to 3.13; using 3.12 syntax keeps them happy
target-version = "py312"

[tool.ruff.lint]
# Common, sane rule set: errors, flake8, imports, bugbear
select = ["E", "F", "I", "B"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = ["src"]
plugins = ["pydantic.mypy"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
diagnosticMode = "workspace"
include = ["src"] 
exclude = ["src/dev_env"]

[tool.ariadne-codegen]
# Where your schema lives, relative to pyproject.toml
schema_path = "src/graphql/schema.graphql"

# Directory with your operations (queries.graphql, mutations.graphql, etc.)
queries_path = "src/graphql/operations"

# Where to generate the client package
# This will create src/graphql_client/...
target_package_name = "graphql_client"
target_package_path = "src"

# Some nice defaults
async_client = true
include_comments = "stable"
convert_to_snake_case = true
